Don't run on PR review
Running on PR review means running again after actions have passed, thus
duplicating CI effort.

We lose the ability to retrigger CI by reviewing PR. This was useful if CI was
flake but we could get it back by either closing and reopening the PR or by
adding a label / comment and triggering CI based on that. Neither of these is
included here.

Unfortunately, there is no way to run additional CI after review but just
before merging. If we need this we'll need to add it via the label trigger (in
which case we need to ensure we don't open ourselves to TOCTOU issues by timing
attacks -- since it takes some time for CI to provision the machines, the
PR author might push new code to the branch. We could fake this via
`workflow_dispacth` trigger, maybe?).

Signed-off-by: Mihai Maruseac <mihai.maruseac@gmail.com>