Remove 'pull_request.closed' trigger
We don't need this since it will be covered by `push` trigger when merging the
PR and is really not needed in all other cases.

This means we also don't need to look at `github.event.pull_reuqest.merged`
state anymore to detect when a closing event is actually due to merging (value
is `"true"` as opposed to `"false"` or missing). Simpler workflows are better.

Signed-off-by: Mihai Maruseac <mihai.maruseac@gmail.com>